Moves queries into lower components
